requirements.txt
dnspython >= 1.16.0
multiping >= 1.1.2